+++ include/navkit/core/math/Types.hpp
@@ -1,0 +1,15 @@
+// Copyright (c) 2026 William Gordon Carter.
+// All Rights Reserved.
+
+#pragma once
+
+#include "navkit/core/config/Types.hpp"
+
+#include <Eigen/Dense>
+
+namespace navkit::core
+{
+
+using Vec3 = Eigen::Matrix<Scalar_t, 3, 1>;
+
+} // namespace navkit::core